admin: split out usage endpoints
